Kamera wychodzi z widoku szczegolow do wlasnego kontrolera
Caly cykl zycia podgladu siedzial w PrinterDetailViewController: rozpoznanie
marki, cztery osobne starty strumienia, lancuch awaryjny RTSP do JPEG dla P1
i A1, dwunastosekundowy limit na pierwsza klatke i sprzatanie. Kamera byla
przez to dostepna z dokladnie jednego miejsca w aplikacji, bo drugie
musialoby to wszystko przepisac.

Teraz jest CameraFeedController: dostaje store i numer seryjny, wystawia
swoj widok i dwie metody, start i stop. Widok szczegolow trzyma jedna jego
instancje i zostawia sobie tylko regule, ktora zna: nie strumieniuj modulu,
ktory uzytkownik schowal.

Przeniesienie jest bez zmian w zachowaniu. CameraView idzie razem z nim,
dostaje tylko sterowany promien zaokraglenia, bo karta w szczegolach chce 10,
a przyszly gospodarz moze chciec ciasniej. Doszla tez jedna wspolna definicja
supportsCamera: wczesniej lista marek z dekodowalnym strumieniem byla wpisana
w warunku widoku szczegolow, a kazda nowa powierzchnia musialaby ja powtorzyc
i mogla sie rozjechac.

Widok szczegolow traci 290 linii i dwa importy, AVFoundation oraz CoreMedia,
ktore byly tam wylacznie dla kamery.

Weryfikacja: swift build, swift build -DGANTRY_LITE, 40 testow Swift, trzy
self-testy, check_ui_parity.py oraz i18n_check.py --strict.
